eks/weather/web/pyproject.toml
[project]
name = "web-ui"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "gradio==5.33.2",
    "boto3==1.38.36",
    "fastapi==0.115.12",
    "itsdangerous==2.2.0",
    "uvicorn==0.34.3",
    "authlib==1.6.0",
    "python-dotenv==1.0.1"
]

[project.scripts]
"web" = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
